+++ lib/src/response/cheermotes_response.dart
@@ -1,7 +1,7 @@
 import 'package:twitch_client/src/response/datum_response.dart';
 
 class CheermotesResponse {
-  final List<Datum> data;
+  final List<CheermotesResponseData> data;
 
   CheermotesResponse({
     required this.data,
@@ -9,6 +9,6 @@ class CheermotesResponse {
 
   factory CheermotesResponse.fromJson(Map<String, dynamic> json) =>
       CheermotesResponse(
-        data: List<Datum>.from(json["data"].map((x) => Datum.fromJson(x))),
+        data: List<CheermotesResponseData>.from(json["data"].map((x) => CheermotesResponseData.fromJson(x))),
       );
 }